backend/pyproject.toml
[project]
name = "change-detector-server"
version = "0.1.0"
description = "FastMCP server for satellite image change detection"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "opencv-python>=4.8.0",
    "pillow>=10.0.0",
    "numpy>=1.24.0,<2.0.0",  # Fix: Constrain to NumPy 1.x for torch compatibility
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "uvicorn>=0.24.0",
    "fastapi>=0.104.0",
    "python-multipart>=0.0.6",
    "aiofiles>=23.0.0",
    "rasterio>=1.3.0",
    "aiohttp>=3.12.14",
    "pydantic>=2.11.7",
    "requests>=2.31.0",

]

[project.optional-dependencies]
# Development dependencies
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]
